+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
  * Fix Parameter '__frch_typeId_0' not found error
  * Fix `UPDATE` statement get user name from `SecurityContextHolder` error
+ * Case order by format from lower hyphen to upper underscore
+ * Fix mybatis flex order query error
 
 ### Optimization